// Polytechnique Montréal/Project 2 - Web Application/client/src/app/declarations/keyboard.enum.ts
export const enum Keyboard {
    A = 'a',
    B = 'b',
    C = 'c',
    D = 'd',
    E = 'e',
    G = 'g',
    I = 'i',
    L = 'l',
    M = 'm',
    O = 'o',
    P = 'p',
    R = 'r',
    S = 's',
    T = 't',
    V = 'v',
    W = 'w',
    X = 'x',
    Z = 'z',
    Digit_1 = '1',
    Digit_2 = '2',
    Digit_3 = '3',
    SHIFT = 'Shift',
    CTRL = 'Control',
    ESC = 'Escape',
    BS = 'Backspace',
    DEL = 'Delete',
    ENT = 'Enter',
    RIGHT_ARROW = 'ArrowRight',
    LEFT_ARROW = 'ArrowLeft',
    UP_ARROW = 'ArrowUp',
    DOWN_ARROW = 'ArrowDown',

    MINUS = '-',
    PLUS = '=',

    F1 = 'F1',
    F2 = 'F2',
    F3 = 'F3',
    F4 = 'F4',
    F5 = 'F5',
    F6 = 'F6',
    F7 = 'F7',
    F8 = 'F8',
    F9 = 'F9',
    F10 = 'F10',
    F11 = 'F11',
    F12 = 'F12',
    CAPS = 'CapsLock',
    ALT = 'Alt',
    NUM = 'NumLock',
    TAB = 'Tab',
    CL = 'Clear',
    END = 'End',
    HOME = 'Home',
    PAGE_UP = 'PageUp',
    PAGE_DOWN = 'PageDown',
}

export const NOT_TEXT_SYMBOLE: string[] = [
    Keyboard.ALT,
    Keyboard.CTRL,
    Keyboard.TAB,
    Keyboard.CAPS,
    Keyboard.SHIFT,
    Keyboard.NUM,
    Keyboard.F1,
    Keyboard.F2,
    Keyboard.F3,
    Keyboard.F4,
    Keyboard.F5,
    Keyboard.F6,
    Keyboard.F7,
    Keyboard.F8,
    Keyboard.F9,
    Keyboard.F10,
    Keyboard.F11,
    Keyboard.F12,
    Keyboard.CL,
    Keyboard.HOME,
    Keyboard.END,
    Keyboard.PAGE_DOWN,
    Keyboard.PAGE_UP,
    Keyboard.BS,
    Keyboard.ENT,
    Keyboard.DEL,
    Keyboard.ESC,
    'Unidentified',
];